backend/requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
numpy==2.0.2
pip==23.0.1
psycopg2-binary==2.9.11
requests==2.32.5
setuptools==79.0.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0
